verifytoken

+++ controllers/genresController.js
@@ -1,7 +1,7 @@
 const Genre = require("../models/Genre");
 
 exports.create = async (req, res) => {
-    // if (req.user.isAdmin) {
+    if (req.user.isAdmin) {
       const newGenre = new Genre(req.body);
       try {
         const savedGenre = await newGenre.save();
@@ -9,13 +9,13 @@ exports.create = async (req, res) => {
       } catch (err) {
         res.status(500).json(err);
       }
-    // } else {
-    //   res.status(403).json("You are not allowed!");
-    // }
+    } else {
+      res.status(403).json("You are not allowed!");
+    }
 }
 
 exports.update = async (req, res) => {
-    // if (req.user.isAdmin) {
+    if (req.user.isAdmin) {
       try {
         const updatedGenre = await Genre.findByIdAndUpdate(
           req.params.id,
@@ -28,22 +28,22 @@ exports.update = async (req, res) => {
       } catch (err) {
         res.status(500).json(err);
       }
-    // } else {
-    //   res.status(403).json("You are not allowed!");
-    // }
+    } else {
+      res.status(403).json("You are not allowed!");
+    }
 }
 
 exports.delete = async (req, res) => {
-    // if (req.user.isAdmin) {
+    if (req.user.isAdmin) {
       try {
         await Genre.findByIdAndDelete(req.params.id);
         res.status(200).json("The genre has been deleted...");
       } catch (err) {
         res.status(500).json(err);
       }
-    // } else {
-    //   res.status(403).json("You are not allowed!");
-    // }
+    } else {
+      res.status(403).json("You are not allowed!");
+    }
 }
 
 exports.getAll = async (req, res) => {
