+++ controllers/seriesController.js
@@ -1,7 +1,7 @@
 const Serie = require("../models/Serie");
 
 exports.create = async (req, res) => {
-    // if (req.user.isAdmin) {
+    if (req.user.isAdmin) {
       const newSerie = new Serie(req.body);
       try {
         const savedSerie = await newSerie.save();
@@ -9,13 +9,13 @@ exports.create = async (req, res) => {
       } catch (err) {
         res.status(500).json(err);
       }
-    // } else {
-    //   res.status(403).json("You are not allowed!");
-    // }
+    } else {
+      res.status(403).json("You are not allowed!");
+    }
 }
 
 exports.update = async (req, res) => {
-    // if (req.user.isAdmin) {
+    if (req.user.isAdmin) {
       try {
         const updatedSerie = await Serie.findByIdAndUpdate(
           req.params.id,
@@ -28,22 +28,22 @@ exports.update = async (req, res) => {
       } catch (err) {
         res.status(500).json(err);
       }
-    // } else {
-    //   res.status(403).json("You are not allowed!");
-    // }
+    } else {
+      res.status(403).json("You are not allowed!");
+    }
 }
 
 exports.delete = async (req, res) => {
-    // if (req.user.isAdmin) {
+    if (req.user.isAdmin) {
       try {
         await Serie.findByIdAndDelete(req.params.id);
         res.status(200).json("The serie has been deleted...");
       } catch (err) {
         res.status(500).json(err);
       }
-    // } else {
-    //   res.status(403).json("You are not allowed!");
-    // }
+    } else {
+      res.status(403).json("You are not allowed!");
+    }
 }
 
 exports.get = async (req, res) => {
@@ -67,14 +67,14 @@ exports.getRandom = async (req, res) => {
 }
 
 exports.getAll = async (req, res) => {
-    // if (req.user.isAdmin) {
+    if (req.user.isAdmin) {
       try {
         const series = await Serie.find();
         res.status(200).json(series.reverse());
       } catch (err) {
         res.status(500).json(err);
       }
-    // } else {
-    //   res.status(403).json("You are not allowed!");
-    // }
+    } else {
+      res.status(403).json("You are not allowed!");
+    }
 }
