+++ controllers/subscribtionsController.js
@@ -1,7 +1,7 @@
 const Subscribtion = require("../models/Subscribtion");
 
 exports.create = async (req, res) => {
-    // if (req.user.isAdmin) {
+    if (req.user.isAdmin) {
       const newSubscribtion = new Subscribtion(req.body);
       try {
         const savedSubscribtion = await newSubscribtion.save();
@@ -9,13 +9,13 @@ exports.create = async (req, res) => {
       } catch (err) {
         res.status(500).json(err);
       }
-    // } else {
-    //   res.status(403).json("You are not allowed!");
-    // }
+    } else {
+      res.status(403).json("You are not allowed!");
+    }
 }
 
 exports.update = async (req, res) => {
-    // if (req.user.isAdmin) {
+    if (req.user.isAdmin) {
       try {
         const updatedSubscribtion = await Subscribtion.findByIdAndUpdate(
           req.params.id,
@@ -28,22 +28,22 @@ exports.update = async (req, res) => {
       } catch (err) {
         res.status(500).json(err);
       }
-    // } else {
-    //   res.status(403).json("You are not allowed!");
-    // }
+    } else {
+      res.status(403).json("You are not allowed!");
+    }
 }
 
 exports.delete = async (req, res) => {
-    // if (req.user.isAdmin) {
+    if (req.user.isAdmin) {
       try {
         await Subscribtion.findByIdAndDelete(req.params.id);
         res.status(200).json("The subscribtion has been deleted...");
       } catch (err) {
         res.status(500).json(err);
       }
-    // } else {
-    //   res.status(403).json("You are not allowed!");
-    // }
+    } else {
+      res.status(403).json("You are not allowed!");
+    }
 }
 
 exports.get = async (req, res) => {
@@ -56,14 +56,14 @@ exports.get = async (req, res) => {
 }
 
 exports.getAll = async (req, res) => {
-    // if (req.user.isAdmin) {
+    if (req.user.isAdmin) {
       try {
         const subscribtions = await Subscribtion.find();
         res.status(200).json(subscribtions.reverse());
       } catch (err) {
         res.status(500).json(err);
       }
-    // } else {
-    //   res.status(403).json("You are not allowed!");
-    // }
+    } else {
+      res.status(403).json("You are not allowed!");
+    }
 }
