+++ controllers/usersController.js
@@ -2,7 +2,7 @@ const User = require("../models/User");
 const CryptoJS = require("crypto-js");
 
 exports.update = async (req, res) => {
-    // if (req.user.id === req.params.id || req.user.isAdmin) {
+    if (req.user.id === req.params.id || req.user.isAdmin) {
       if (req.body.password) {
         req.body.password = CryptoJS.AES.encrypt(
           req.body.password,
@@ -22,22 +22,22 @@ exports.update = async (req, res) => {
       } catch (err) {
         res.status(500).json(err);
       }
-    // } else {
-    //   res.status(403).json("You can update only your account!");
-    // }
+    } else {
+      res.status(403).json("You can update only your account!");
+    }
 }
 
 exports.delete = async (req, res) => {
-    // if (req.user.id === req.params.id || req.user.isAdmin) {
+    if (req.user.id === req.params.id || req.user.isAdmin) {
       try {
         await User.findByIdAndDelete(req.params.id);
         res.status(200).json("User has been deleted...");
       } catch (err) {
         res.status(500).json(err);
       }
-    // } else {
-    //   res.status(403).json("You can delete only your account!");
-    // }
+    } else {
+      res.status(403).json("You can delete only your account!");
+    }
 }
 
 exports.get = async (req, res) => {
@@ -62,7 +62,7 @@ exports.getWichlist = async (req, res) => {
 
 exports.getAll = async (req, res) => {
     const query = req.query.new;
-    // if (req.user.isAdmin) {
+    if (req.user.isAdmin) {
       try {
         const users = query
           ? await User.find().sort({ _id: -1 }).limit(5)
@@ -71,9 +71,9 @@ exports.getAll = async (req, res) => {
       } catch (err) {
         res.status(500).json(err);
       }
-    // } else {
-    //   res.status(403).json("You are not allowed to see all users!");
-    // }
+    } else {
+      res.status(403).json("You are not allowed to see all users!");
+    }
 }
 
 exports.getStats = async (req, res) => {
